+++ src/containers/indexContainer.js
@@ -1,5 +1,5 @@
 import React from "react";
-import NavigationBar from "../Layouts/Navigationbar";
+import NavigationBar from "../components/Navigationbar";
 import Typography from "@material-ui/core/Typography";
 import Grid from "../components/homepage/grid";
 import WhatWeDo from "../components/homepage/whatwedo";
@@ -32,72 +32,68 @@ const section = {
   marginBottom: `10%`
 };
 
-class IndexContainer extends React.Component {
-  render() {
-    return (
-      <div>
-        <NavigationBar />
-        <img src="index.jpg" style={styles} />
+export default () => {
+  return (
+    <div>
+      <NavigationBar />
+      <img src="index.jpg" style={styles} />
+      <Typography
+        component="h2"
+        align="center"
+        variant="h5"
+        gutterBottom
+        style={headingText}
+      >
+        We are Zairza
+      </Typography>
+      <hr style={hrStyle} />
+      <div class="container">
         <Typography
           component="h2"
-          align="center"
+          align="justify"
           variant="h5"
           gutterBottom
-          style={headingText}
+          style={innerPara}
         >
-          We are Zairza
+          ZAIRZA is a cult for CETians who want to be ahead in the game. Sure we
+          all have talent, in this club we put our talents through many
+          obstacles and we conquer them. We walk out as winners who not only
+          read books all these years but test the theories and sometimes make
+          our own.
         </Typography>
-        <hr style={hrStyle} />
-        <div class="container">
-          <Typography
-            component="h2"
-            align="justify"
-            variant="h5"
-            gutterBottom
-            style={innerPara}
-          >
-            ZAIRZA is a cult for CETians who want to be ahead in the game. Sure
-            we all have talent, in this club we put our talents through many
-            obstacles and we conquer them. We walk out as winners who not only
-            read books all these years but test the theories and sometimes make
-            our own.
-          </Typography>
-        </div>
-        <div className="container-fluid" style={section}>
-          <Grid />
-        </div>
-        <Typography
-          component="h2"
-          align="center"
-          variant="h5"
-          gutterBottom
-          style={headingText}
-        >
-          What We Do
-        </Typography>
-        <hr style={hrStyle} />
-        <div class="container">
-          <Typography
-            component="h2"
-            align="justify"
-            variant="h5"
-            gutterBottom
-            style={innerPara}
-          >
-            Extensive research and development is carried out which equips the
-            members with knowledge on various technological fields. Members are
-            actively engaged in taking the society to greater height through
-            their continuous efforts in the field of Engineering. The legacy of
-            Zairza flows down through each batch of CET which empowers us to
-            stand ahead of our counterparts.
-          </Typography>
-        </div>
-        <div className="container" style={section}>
-          <WhatWeDo />
-        </div>
       </div>
-    );
-  }
-}
-
-export default IndexContainer;
+      <div className="container-fluid" style={section}>
+        <Grid />
+      </div>
+      <Typography
+        component="h2"
+        align="center"
+        variant="h5"
+        gutterBottom
+        style={headingText}
+      >
+        What We Do
+      </Typography>
+      <hr style={hrStyle} />
+      <div class="container">
+        <Typography
+          component="h2"
+          align="justify"
+          variant="h5"
+          gutterBottom
+          style={innerPara}
+        >
+          Extensive research and development is carried out which equips the
+          members with knowledge on various technological fields. Members are
+          actively engaged in taking the society to greater height through their
+          continuous efforts in the field of Engineering. The legacy of Zairza
+          flows down through each batch of CET which empowers us to stand ahead
+          of our counterparts.
+        </Typography>
+      </div>
+      <div className="container" style={section}>
+        <WhatWeDo />
+      </div>
+    </div>
+  );
+};
