+++ src/pages/alumni.js
@@ -1,6 +1,6 @@
 import React from "react";
 import Link from "gatsby-link";
-import NavigationBar from "../Layouts/Navigationbar";
+import NavigationBar from "../components/Navigationbar";
 
 export default () => (
   <div>
