+++ src/pages/index.js
@@ -1,10 +1,6 @@
-import React from "react";
+import React, { Component } from "react";
 import IndexContainer from "../containers/indexContainer";
 
-class Index extends React.Component {
-  render() {
-    return <IndexContainer />;
-  }
-}
-
-export default Index;
+export default () => {
+  return <IndexContainer />;
+};
